Move passport configuration into its own file
